--- src/components/Ui/Button/index.tsx
@@ -1,7 +1,0 @@
-import styled from "styled-components"
-
-export const Button = styled.button`
-  padding: 10px 20px;
-  border-radius: 5px;
-  color: green;
-`
