--- src/components/Ui/Container/index.tsx
@@ -1,9 +1,0 @@
-import styled from "styled-components"
-
-export const Container = styled.section`
-  grid-column: LS / RS 1;
-  padding: 60px 0;
-  @media (max-width: 768px) {
-    padding: 30px 0;
-  }
-`
